refactor: split file
Split utilities file into mock-work-dir and mock-console
