+++ test/mock-work-dir.ts
@@ -1,19 +1,16 @@
-import fse from "fs-extra";
 import path from "path";
 import os from "os";
-import testConsole from "test-console";
 import { PkgManifest } from "../src/types/global";
+import fse from "fs-extra";
 import _ from "lodash";
 
 export type ManifestCreationOptions = {
   manifest: boolean | PkgManifest;
   editorVersion?: string;
 };
-
 export const getWorkDir = function (pathToTmp: string): string {
   return path.join(os.tmpdir(), pathToTmp);
 };
-
 export const createWorkDir = function (
   pathToTmp: string,
   { manifest, editorVersion }: ManifestCreationOptions
@@ -37,30 +34,7 @@ export const createWorkDir = function (
     );
   }
 };
-
 export const removeWorkDir = function (pathToTmp: string) {
   const cwd = getWorkDir(pathToTmp);
   fse.removeSync(cwd);
-};
-
-export const getOutputs = function (
-  stdouInspect: testConsole.Inspector,
-  stderrInsepct: testConsole.Inspector
-): [string, string] {
-  const results: [string, string] = [
-    stdouInspect.output.join(""),
-    stderrInsepct.output.join(""),
-  ];
-  stdouInspect.restore();
-  stderrInsepct.restore();
-  return results;
-};
-
-export const getInspects = function (): [
-  testConsole.Inspector,
-  testConsole.Inspector
-] {
-  const stdoutInspect = testConsole.stdout.inspect();
-  const stderrInspect = testConsole.stderr.inspect();
-  return [stdoutInspect, stderrInspect];
 };
